requirements.txt
# Core dependencies
pydantic>=2.0.0
PyYAML>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
pylint>=3.0.0
black>=23.0.0
